Modified the Autoreleasepool data structure
Instead of using an std::linked list to keep track of our pending releases a
custom linked list might be better suited

Autoreleasepools have a unique use case in that we want fast insertion on one
end, once we drain we want to empty everything as quickly as possible.

Linked lists work great for fast insertion but the problem is we'll be
constantly hitting cache misses. Taking advantage of spatial locality of the
traditional array would be nice.

The design here is we have a stack of ReferenceCounted objects being maintained
in fixed size buffers.

HEAD -> [buffer of X size] -> [buffer of X size] -> nullptr

Once a single autorelease buffer has been filled up, it adds another buffer at
the front of the linked list. When it comes time to drain it will start at the
front of the linked list (most recently added object) and work towards the
oldest object.

We go in this order to prevent cache thrashing. If we released the oldest object
first and worked forward we could be evicting some of our more recent buffers
out of the cache that we'd just need to load back in anyway. Working backwards
takes advantage of cache temporal locality.